+++ jeap-spring-boot-security-client-starter/src/main/java/ch/admin/bit/jeap/security/client/OAuth2ClientRegistryPostprocessor.java
@@ -12,18 +12,19 @@ import java.util.Set;
  * This bean postprocessor implements a default for the scope configuration properties of clients in the Spring Boot
  * client registry. It replaces an empty scope configuration with the scope 'openid', which is mandatory for OIDC auth flows.
  * Defaulting the scope property to 'openid' makes it unnecessary to repeatedly configure such a scope for every client.
- *
+ * <p>
  * This bean postprocessor has been added because of a a problem with the current Spring Security implementation. If an
  * OAuth2 client does not specify a scope in its client configuration, Spring Security will add certain scopes to its requests
  * to the authorization server on its own. More precisely, Spring Security will add all scopes that the authorization server
  * lists at its configuration endpoint e.g. at issuer/.well-known/openid-configuration. This does not make sense, especially
  * if those scopes include the scope 'offline_access', which makes the authorization server needlessly accumulate
- * long lived refresh tokens.
- *
+ * long-lived refresh tokens.
+ * <p>
  * Spring Security 5.4 changes the scope handling by no longer fetching scopes from the authorization server's configuration
  * endpoint and leaving an empty scope configuration as is. However, this might cause problems for OIDC (but not for 'pure' OAuth2),
- * because of the missing mandatory 'openid' scope. See https://github.com/spring-projects/spring-security/issues/8514 for details.
- * When upgrading the jeap-spring-boot-startes to a newer Spring Boot version with a newer Spring Security version than the
+ * because of the missing mandatory 'openid' scope. See <a href="https://github.com/spring-projects/spring-security/issues/8514">
+ * issues 8514</a> for details.
+ * When upgrading the jeap-spring-boot-starters to a newer Spring Boot version with a newer Spring Security version than the
  * current 5.3 version, we might want to reconsider the situation.
  */
 @Slf4j
@@ -33,8 +34,7 @@ public class OAuth2ClientRegistryPostprocessor implements BeanPostProcessor {
     private static final String OPENID_SCOPE = "openid";
 
     public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
-        if (bean instanceof OAuth2ClientProperties) {
-            OAuth2ClientProperties properties = (OAuth2ClientProperties) bean;
+        if (bean instanceof OAuth2ClientProperties properties) {
             properties.getRegistration().forEach(this::defaultToOpenidScopeForEmptyScope);
         }
         return bean;