+++ jeap-spring-boot-security-client-starter/src/main/java/ch/admin/bit/jeap/security/restclient/OAuth2RestClientConfiguration.java
@@ -4,34 +4,30 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.boot.autoconfigure.security.oauth2.client.ClientsConfiguredCondition;
+import org.springframework.boot.autoconfigure.web.client.RestClientAutoConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
 import org.springframework.security.oauth2.client.*;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
 import org.springframework.web.client.RestClient;
 
 /**
  * This configuration makes a RestClient builder factory available that can create a RestClient.Builder instance
- * that builds RestClient instances that automatically add an OAuth2 access token as bearer to the RestClient exchanges.
+ * that builds RestClient instances that automatically add an OAuth2 access token as a bearer token to the RestClient exchanges.
  */
 @Slf4j
 @AutoConfiguration
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-@ConditionalOnClass(RestClient.class)
+@ConditionalOnBean(RestClient.Builder.class)
+@AutoConfigureAfter(RestClientAutoConfiguration.class)
 @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
 public class OAuth2RestClientConfiguration {
 
     @Configuration
-    @Order(Ordered.LOWEST_PRECEDENCE -1) // to be executed before ServletRestClientForNoOAuthClientsConfiguration
-    @Conditional(ClientsConfiguredCondition.class)
-    @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
+    @ConditionalOnBean({ClientRegistrationRepository.class, OAuth2AuthorizedClientService.class})
     public static class ServletRestClientForOAuthClientsConfiguration {
         @Bean
         public JeapOAuth2RestClientBuilderFactory jeapOAuth2RestclientBuilderFactory(RestClient.Builder builder, ClientRegistrationRepository clientRegistrationRepository, OAuth2AuthorizedClientService clientService) {
@@ -49,14 +45,12 @@ public class OAuth2RestClientConfiguration {
     }
 
     @Configuration
-    @ConditionalOnMissingBean(JeapOAuth2RestClientBuilderFactory.class)
-    @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
+    @ConditionalOnMissingBean({ClientRegistrationRepository.class, OAuth2AuthorizedClientService.class, JeapOAuth2RestClientBuilderFactory.class})
     public static class ServletRestClientForNoOAuthClientsConfiguration {
         @Bean
-        public JeapOAuth2RestClientBuilderFactory jeapOAuth2RestclientBuilderFactory(RestClient.Builder builder) {
+        public JeapOAuth2RestClientBuilderFactory jeapOAuth2RestClientBuilderFactory(RestClient.Builder builder) {
             return new DefaultJeapOAuth2RestClientBuilderFactory(builder, null, null);
         }
     }
-
 
 }