+++ jeap-spring-boot-security-starter-it-webmvc/src/test/java/ch/admin/bit/jeap/security/it/client/webmvc/OAuth2WebClientWebMvcIT.java
@@ -1,8 +1,9 @@
 package ch.admin.bit.jeap.security.it.client.webmvc;
 
 import ch.admin.bit.jeap.security.client.JeapOAuth2WebclientBuilderFactory;
+import ch.admin.bit.jeap.security.it.bearertoken.BearerTokenTestExtension;
+import ch.admin.bit.jeap.security.it.bearertoken.BearerTokenUrl;
 import ch.admin.bit.jeap.security.it.mockserver.OAuth2MockServer;
-import ch.admin.bit.jeap.security.it.resource.BearerTokenResource;
 import ch.admin.bit.jeap.security.it.resource.OAuth2TestGateway;
 import ch.admin.bit.jeap.security.it.resource.OAuth2TestGateway.WebClientTokenSource;
 import ch.admin.bit.jeap.security.resource.token.JeapAuthenticationContext;
@@ -11,13 +12,13 @@ import ch.admin.bit.jeap.security.test.resource.JeapAuthenticationTestTokenBuild
 import ch.admin.bit.jeap.security.test.resource.extension.WithAuthentication;
 import com.github.tomakehurst.wiremock.matching.StringValuePattern;
 import com.nimbusds.jwt.JWT;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
@@ -31,7 +32,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @ActiveProfiles("client")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@ExtendWith(BearerTokenTestExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class OAuth2WebClientWebMvcIT {
 
     private static final int OAUTH_MOCK_SERVER_PORT = 9090;
@@ -42,14 +44,28 @@ public class OAuth2WebClientWebMvcIT {
     private static final String DEFAULT_SCOPE = "openid";
     private static final String USER_DEFINED_SCOPE = "user-defined-scope";
 
-    private static OAuth2MockServer oauth2MockServer = new OAuth2MockServer(OAUTH_MOCK_SERVER_PORT, OAUTH_MOCK_SERVER_BASE_PATH);
+    private static final OAuth2MockServer oauth2MockServer = new OAuth2MockServer(OAUTH_MOCK_SERVER_PORT, OAUTH_MOCK_SERVER_BASE_PATH);
+
+    private static String bearerTokenUrl;
+
+    @LocalServerPort
+    int serverPort;
+
+    @Autowired
+    @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
+    JeapOAuth2WebclientBuilderFactory jeapOAuth2WebclientBuilderFactory;
 
     static {
         oauth2MockServer.start();
     }
 
-    @AfterEach
-    void reset() {
+    @BeforeAll
+    static void init(@BearerTokenUrl String url) {
+        bearerTokenUrl = url;
+    }
+
+    @BeforeEach
+    void setup() {
         oauth2MockServer.reset();
     }
 
@@ -58,15 +74,19 @@ public class OAuth2WebClientWebMvcIT {
         oauth2MockServer.stop();
     }
 
-    @Value("${server.port}")
-    int serverPort;
-
-    @Autowired
-    @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
-    JeapOAuth2WebclientBuilderFactory jeapOAuth2WebclientBuilderFactory;
-
     @Test
     void testCreateForClientId_WhenOutsideOfAnHttpRequest_ThenTokenFetchedForClient() {
+        final JWT token = oauth2MockServer.stubTokenRequestWithNearlyExpiredToken();
+
+        final WebClient webClient = jeapOAuth2WebclientBuilderFactory.createForClientRegistryId(TEST_CLIENT_ID).build();
+        final String response = callResource(webClient);
+
+        oauth2MockServer.verifyTokenRequest(1);
+        assertThat(response).isEqualTo(token.serialize());
+    }
+
+    @Test
+    void testCreateForClientId_WhenOutsideOfAnHttpRequest_ThenTokenFetchedForClient_2() {
         final JWT token = oauth2MockServer.stubTokenRequestWithNearlyExpiredToken();
 
         final WebClient webClient = jeapOAuth2WebclientBuilderFactory.createForClientRegistryId(TEST_CLIENT_ID).build();
@@ -216,7 +236,7 @@ public class OAuth2WebClientWebMvcIT {
 
     private String callResource(WebClient webClient) {
         return webClient.get().
-                uri("http://localhost:" + serverPort + BearerTokenResource.API_PATH).
+                uri(bearerTokenUrl).
                 retrieve().
                 bodyToMono(String.class).
                 block(Duration.ofSeconds(10));
@@ -235,4 +255,13 @@ public class OAuth2WebClientWebMvcIT {
     private JeapAuthenticationToken defaultAuthentication() {
         return  JeapAuthenticationTestTokenBuilder.create().withContext(JeapAuthenticationContext.SYS).build();
     }
+
+    @TestConfiguration
+    static class GatewayConfig {
+        @Bean
+        OAuth2TestGateway oAuth2TestGateway(JeapOAuth2WebclientBuilderFactory jeapOAuth2WebclientBuilderFactory) {
+            return new OAuth2TestGateway(jeapOAuth2WebclientBuilderFactory, bearerTokenUrl);
+        }
+    }
+
 }