+++ jeap-spring-boot-security-starter-it/src/main/java/ch/admin/bit/jeap/security/it/bearertoken/BearerTokenUrl.java
@@ -1,0 +1,9 @@
+package ch.admin.bit.jeap.security.it.bearertoken;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD, ElementType.PARAMETER})
+@Documented
+public @interface BearerTokenUrl {
+}